react-python-lab/requirements.txt
pywebview
cryptography
pyinstaller

# query diff
oracledb
psycopg2-binary
pandas

# api client
requests

# screenshot
pillow 
pywin32
pyautogui